# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/surveys/investor": {
      "post": {
        "operationId": "SurveyController_investorCompleteSurvey",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InvestorSurveyDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "surveys"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/surveys/business-owner": {
      "post": {
        "operationId": "SurveyController_businessOwnerCompleteSurvey",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BusinessOwnerSurveyDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "surveys"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Asabik API Docs",
    "description": "All endpoints",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "InvestorSurveyQuestionDto": {
        "type": "object",
        "properties": {
          "question": {
            "type": "string"
          },
          "answer": {
            "type": "boolean"
          }
        },
        "required": [
          "question",
          "answer"
        ]
      },
      "InvestorSurveyDto": {
        "type": "object",
        "properties": {
          "questions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/InvestorSurveyQuestionDto"
            }
          },
          "fullName": {
            "type": "string"
          },
          "companyName": {
            "type": "string"
          },
          "website": {
            "type": "string"
          }
        },
        "required": [
          "questions",
          "fullName"
        ]
      },
      "CreateInvestmentRequestDto": {
        "type": "object",
        "properties": {
          "purposeOfTheLoan": {
            "type": "string",
            "enum": [
              "Working capital",
              "Equipment Purchase",
              "Inventory Purchase",
              "Expand Business",
              "Finance Debt Obligation",
              "Marketing or Design",
              "Emergency",
              "Other"
            ]
          },
          "requiredCapital": {
            "type": "number"
          },
          "returnTerm": {
            "type": "number"
          },
          "netReturn": {
            "type": "number"
          },
          "netReturnToShare": {
            "type": "number"
          },
          "loanPurpose": {
            "type": "string"
          },
          "helpIncreaseProfit": {
            "type": "string"
          },
          "profitIncrease": {
            "type": "number"
          }
        },
        "required": [
          "purposeOfTheLoan",
          "requiredCapital",
          "returnTerm",
          "netReturn",
          "netReturnToShare",
          "loanPurpose",
          "helpIncreaseProfit",
          "profitIncrease"
        ]
      },
      "BusinessOwnerSurveyDto": {
        "type": "object",
        "properties": {
          "businessStartDate": {
            "type": "string"
          },
          "investmentRequest": {
            "$ref": "#/components/schemas/CreateInvestmentRequestDto"
          },
          "businessStructure": {
            "type": "string",
            "enum": [
              "Corporation",
              "LLC",
              "Sole proprietorship",
              "Partnership",
              "Single-member LLC",
              "Nonprofit organization"
            ]
          },
          "avrMonthlySales": {
            "type": "number"
          },
          "avrMonthlyNetProfit": {
            "type": "number"
          },
          "totalLastYearNetProfit": {
            "type": "number"
          },
          "employeesNo": {
            "type": "number"
          },
          "businessSubindustryId": {
            "type": "number"
          },
          "website": {
            "type": "string"
          }
        },
        "required": [
          "businessStartDate",
          "investmentRequest",
          "businessStructure",
          "avrMonthlySales",
          "avrMonthlyNetProfit",
          "totalLastYearNetProfit",
          "employeesNo",
          "businessSubindustryId",
          "website"
        ]
      },
...
    }
  }
}